requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
pylint
python-dotenv==1.0.0
django-filter
pandas
openpyxl